Revert "Added ssl_mode flag to mysqldb"
This reverts commit 2e4dec934cb5215d628e02ed717454c165a33e4d.

I approved this merge before observing that the "ssl_mode"
flag is not actually part of the separate "ssl" dictionary,
so this patch as given is unnecessary as we already pass
all strings into opts.  but also I can't get the flag to work
on either mysqlclient or pymysql, it's not recognized.  so
I need specific info from the submitter what it is that is
actually needed, and if a change is needed it has to be
appropriate for the drivers in use.