Ensure the "orm" plugin is used unconditionally for bundles
This also introduces that the "orm" plugin may be used
when the plugin_subject is None.

Fixed regression where the :paramref:`.Bundle.single_entity` flag would
take effect for a :class:`.Bundle` even though it were not set.
Additionally, this flag is legacy as it only makes sense for the
:class:`_orm.Query` object and not 2.0 style execution.  a deprecation
warning is emitted when used with new-style execution.

Fixes: #5702
Change-Id: If9f43365b9d966cb398890aeba2efa555658a7e5